+++ sigle/src/pages/[username].page.tsx
@@ -10,6 +10,7 @@ import { PublicHome } from '../modules/publicHome';
 import { sigleConfig } from '../config';
 import { StoryFile, SettingsFile } from '../types';
 import Error from './_error.page';
+import { fetchPublicStories, fetchSettings } from '../utils/gaia/fetch';
 
 interface PublicHomePageProps {
   statusCode: number | boolean;
@@ -32,36 +33,6 @@ export const PublicHomePage: NextPage<PublicHomePageProps> = ({
 
   // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
   return <PublicHome file={file!} settings={settings!} userInfo={userInfo!} />;
-};
-
-export const fetchPublicStories = async (bucketUrl: string) => {
-  let file;
-  let statusCode: false | number = false;
-  const data = await fetch(`${bucketUrl}publicStories.json`);
-  if (data.status === 200) {
-    file = await data.json();
-  } else if (data.status === 404) {
-    // If file is not found we set an empty array to show an empty list
-    file = { stories: [] };
-  } else {
-    statusCode = data.status;
-  }
-  return { file, statusCode };
-};
-
-export const fetchSettings = async (bucketUrl: string) => {
-  let file;
-  let statusCode: false | number = false;
-  const data = await fetch(`${bucketUrl}settings.json`);
-  if (data.status === 200) {
-    file = await data.json();
-  } else if (data.status === 404) {
-    // If file is not found we set an empty object
-    file = {};
-  } else {
-    statusCode = data.status;
-  }
-  return { file, statusCode };
 };
 
 export const getServerSideProps: GetServerSideProps<
